requirements.txt
rembg
pillow
opencv-python
numpy
pyinstaller
onnxruntime
